Modify the test suite to take less time.
This commit should address the user requests in #9. It does so by greatly
reducing the number of keys generated in the course of testing. Instead of
generating a fresh key pair for every QuickCheck property that requires a
key pair, it pre-generates a small number of key pairs from which the
properties randomly select one. This should be a reasonable trade-off
between runtime of the test executable and coverage of a wide variety of
possible test inputs.