Add a test case example to show how to get around #19.
The core problem we have with issue #19 is that we don't actually process
the ASN.1 describing the hash function used. If we did, we could unwind
this issue on our own ... maybe. There is a core issue here that, right
now, the library is structured so that it doesn't need to know about all
the hash functions the user might want. (Although it provides links to
SHA for convenience.) Maintaining this ability while also properly handling
ASN.1 case would likely require some state ...

That being said, I don't know that I have any plans to add proper ASN.1
parsing to this library. Thus, this mechanism -- providing a specific hash
function definition for this case -- is probably the only way forward.